Friendbot, comments: Extract Serve method and use in friendbot and bifrost, add router to friendbot
