+++ support/http/server/chi-router.go
@@ -1,10 +1,21 @@
 package server
 
 import (
+	"fmt"
 	"net/http"
+	"time"
 
 	"github.com/go-chi/chi"
+	"github.com/stellar/go/support/log"
+	"github.com/tylerb/graceful"
+	"golang.org/x/net/http2"
 )
+
+// TLSConfig specifies the TLS portion of a config
+type TLSConfig struct {
+	CertificateFile string `toml:"certificate-file" valid:"required"`
+	PrivateKeyFile  string `toml:"private-key-file" valid:"required"`
+}
 
 // routeMaker translates the HTTP string method to the chi-equivalent route operation
 var routeMaker = map[string]func(*chi.Mux, string, http.HandlerFunc){
@@ -32,4 +43,41 @@ func NewRouter(c *Config) *chi.Mux {
 	}
 
 	return mux
+}
+
+// Serve starts a web server by binding it to a socket and setting up the shutdown signals
+func Serve(router *chi.Mux, port int, tls *TLSConfig) {
+	http.Handle("/", router)
+
+	addr := fmt.Sprintf(":%d", port)
+
+	srv := &graceful.Server{
+		Timeout: 10 * time.Second,
+
+		Server: &http.Server{
+			Addr:    addr,
+			Handler: http.DefaultServeMux,
+		},
+
+		ShutdownInitiated: func() {
+			log.Info("received signal, gracefully stopping")
+		},
+	}
+
+	http2.ConfigureServer(srv.Server, nil)
+
+	log.Info("Starting web service on " + addr)
+
+	var err error
+	if tls != nil {
+		err = srv.ListenAndServeTLS(tls.CertificateFile, tls.PrivateKeyFile)
+	} else {
+		err = srv.ListenAndServe()
+	}
+
+	if err != nil {
+		log.Panic(err)
+	}
+
+	log.Info("stopped")
 }
